requirement.txt
ansys-corba              0.1.1
ansys-grpc-mapdl         0.2.0
ansys-mapdl-core         0.58.1
ansys-mapdl-reader       0.50.6
appdirs                  1.4.4
cachetools               4.2.1
certifi                  2020.12.5
chardet                  4.0.0
cycler                   0.10.0
geomdl                   5.3.1
google-api-core          1.26.1
google-api-python-client 2.0.2
google-auth              1.27.1
google-auth-httplib2     0.1.0
googleapis-common-protos 1.53.0
grpcio                   1.36.1
httplib2                 0.19.0
idna                     2.10
imageio                  2.9.0
joblib                   1.0.1
kiwisolver               1.3.1
matplotlib               3.3.4
meshio                   4.3.11
numpy                    1.20.1
packaging                20.9
pandas                   1.2.3
Pillow                   8.1.2
pip                      21.0.1
protobuf                 3.15.5
pyasn1                   0.4.8
pyasn1-modules           0.2.8
pyiges                   0.1.5
pyparsing                2.4.7
PyQt5                    5.15.4
PyQt5-Qt5                5.15.2
PyQt5-sip                12.8.1
python-dateutil          2.8.1
pytz                     2021.1
pyvista                  0.29.0
requests                 2.25.1
rsa                      4.7.2
scikit-learn             0.24.1
scipy                    1.6.1
scooby                   0.5.6
setuptools               54.1.1
six                      1.15.0
sklearn                  0.0
threadpoolctl            2.1.0
tqdm                     4.59.0
transforms3d             0.3.1
uritemplate              3.0.1
urllib3                  1.26.3
vtk                      9.0.1
